Allow to update search stat without outputing mess
Setting g:anzu_enable_CursorHold_AnzuUpdateSearchStatus to 2 will only
update the cache, but not output anything. This is useful, if you want
the statistics only to be output by the statusline and not output on the
command line.

In order to do that, refactor the Cursor Moved autocommand and put its
logic into a new function, so it is easier to understand.