do not reset curswant after s:searchpos() calls
when s:searchpos() was called after a cursormoved event, this would
reset the actually wanted column. Take this example:

,----
| a long line with many text characters!
| a shorter line
| along line with many text characters!
`----

Put the cursor on the '!' of the first line, replace by '.' using 'r.'
now move down two lines: '2j'
Expected result: cursor is on the '!' of the third line.
Actual result: cursor is on the t of with

This can be prevented by using winsaveview(), which among other things,
does reset the curswant attribute to the correct value.